+++ build.gradle
@@ -24,23 +24,15 @@ repositories {
     }
 }
 
-dependencyManagement {
-    imports {
-        mavenBom(releaseMode ? 'com.epam.reportportal:commons-bom:' + getProperty('bom.version') : 'com.epam.reportportal:commons-bom:5.14.2')
-    }
-}
-
 ext['spring-boot.version'] = "${springBootVersion}"
 
 dependencies {
     if (releaseMode) {
-        implementation 'com.epam.reportportal:commons-dao'
-        implementation 'com.epam.reportportal:plugin-api'
-        annotationProcessor 'com.epam.reportportal:plugin-api'
+        implementation 'com.epam.reportportal:service-api'
+        annotationProcessor 'com.epam.reportportal:service-api'
     } else {
-        implementation 'com.github.reportportal:commons-dao:ae1f509'
-        implementation 'com.github.reportportal:plugin-api:e1786c4'
-        annotationProcessor 'com.github.reportportal:plugin-api:e1786c4'
+        implementation 'com.github.reportportal:service-api:3e34a90'
+        annotationProcessor 'com.github.reportportal:service-api:3e34a90'
     }
     implementation "io.swagger.core.v3:swagger-annotations-jakarta:2.2.27"
     implementation 'com.squareup.okhttp:okhttp:2.7.5'
