+++ gradle.properties
@@ -2,5 +2,5 @@ version=5.13.0
 description=EPAM Report Portal. Azure DevOps plugin
 pluginId = Azure DevOps
 
-springBootVersion=3.4.2
-lombokVersion=1.18.36
+springBootVersion=3.4.11
+lombokVersion=1.18.42
