+++ src/main/java/com/epam/reportportal/extension/azure/command/role/ProjectMemberCommand.java
@@ -4,16 +4,16 @@ import static com.epam.reportportal.extension.azure.command.utils.CommandParamUt
 import static com.epam.reportportal.extension.azure.command.utils.CommandParamUtils.retrieveLong;
 import static java.util.Optional.ofNullable;
 
-import com.epam.reportportal.rules.commons.validation.BusinessRule;
-import com.epam.reportportal.rules.exception.ErrorType;
-import com.epam.reportportal.rules.exception.ReportPortalException;
-import com.epam.ta.reportportal.commons.ReportPortalUser;
-import com.epam.ta.reportportal.dao.ProjectRepository;
-import com.epam.ta.reportportal.dao.organization.OrganizationRepositoryCustom;
-import com.epam.ta.reportportal.entity.organization.Organization;
-import com.epam.ta.reportportal.entity.organization.OrganizationRole;
-import com.epam.ta.reportportal.entity.project.Project;
-import com.epam.ta.reportportal.entity.user.UserRole;
+import com.epam.reportportal.infrastructure.persistence.commons.ReportPortalUser;
+import com.epam.reportportal.infrastructure.rules.commons.validation.BusinessRule;
+import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
+import com.epam.reportportal.infrastructure.persistence.dao.ProjectRepository;
+import com.epam.reportportal.infrastructure.persistence.dao.organization.OrganizationRepositoryCustom;
+import com.epam.reportportal.infrastructure.persistence.entity.organization.Organization;
+import com.epam.reportportal.infrastructure.persistence.entity.organization.OrganizationRole;
+import com.epam.reportportal.infrastructure.persistence.entity.project.Project;
+import com.epam.reportportal.infrastructure.persistence.entity.user.UserRole;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
