+++ src/main/java/com/epam/reportportal/extension/azure/info/PluginInfoProvider.java
@@ -1,6 +1,6 @@
 package com.epam.reportportal.extension.azure.info;
 
-import com.epam.ta.reportportal.entity.integration.IntegrationType;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationType;
 
 /**
  * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
